Fix flaky CommonPreviewRepresentationTest
There were two causes for the flakiness:

1. Some of the tests were not de-activating the preview properly at the
end of the test. When the following test was requesting builds, the
build listeners of the previous representations were sending refresh
requests that sometimes would end up
interfering with the new test.

2. When a refresh request was cancelled with invalidate=true, we would
re-invalidate the preview representation in an invokeOnCompletion.
However, it turns out the invokeOnCompletion is run concurrently to the
new request that triggered the cancellation. In some cases, the new
request would set invalidate to false before the invokeOnCompletion
would reset the invalidate to true. This led to a timeout in the tests
as we were waiting for invalidate to be false.

Fixes: 379138169
Test: CommonPreviewRepresentationTest x1000
Change-Id: I472e1cf50b05890cd5240d253a6ddec0faf7bc87

GitOrigin-RevId: 5c5ff0a802c1ce3437debc5d514d2bdc6766f7a7