More closely control Manifest light class generation
This change creates a new ManifestClassToken that hides
gradle-specific logic around when and how to create Manifest light
classes

This is under the assumption that we do not ever want to create
Manifest light classes for androidTest. If we want to do that we will
also need to use the abstraction of this logic to decide which package
name to use for a module's Manifest light class and whether to
generate that module's Manifest light class at all

While this addresses the immediate bug we're encountering, work
remains to tie in the Gradle property android.generateManifestClass.

Bug: 380110848
Test: Not yet, will add tests once approach is greenlit
Change-Id: I4e6c35be849e61aa07d79af1a90b24bf9c520299

GitOrigin-RevId: a0ac06f0134d47899299914d76ebffed5b3751d0