Update entry point for Preview toolbar actions
Replace the auto-hide toolbar with a single "More actions" popup
action. Also, add the toolbar actions to the right-click context
menu of each preview as an alternative entry point. For that change,
we also needed to pass the SceneViewPeerPanel as the target component
of the right-click handler for the right DataContext to be used when
updating and activating the actions.

Bug: 378569700
Test: ComposePreviewTest, RenderErrorTest
Change-Id: I7946b0b3b264a6cf3a3458371bb7630d89d1e3a9

GitOrigin-RevId: 5712b3a00369e8c91609281733d9a935593b43e6