Be more selective when listening to Psi events in GradleFiles
It's possible for PsiFile objects to be created for files that aren't
really associated with the project: that is, just because
psiFile.project is equal to the project, does not mean that the actual
file associated with the psiFile is in any way connected to the
project; instead, something else might simply have requested that
PsiFile object from a PsiManager and a VirtualFile.

That something else might have been GradleFiles itself; it deals with
adding or removing a PsiTreeListener, and in the process constructs a
psiFile without checking whether the underlying virtualFile is
relevant.  Stop doing that.

However, that's not enough in general (and manual testing suggests in
particular), so also defend against PsiTreeChangedEvents associated
with PsiFiles that are not relevant to the current project similarly.

We use a DirectoryScope, rather than a ProjectScope, because we are
interested in changes in files under <projectRoot>/.gradle, which is
explicitly excluded from the module content roots under the Gradle
plugin.  Using a DirectoryScope is potentially wrong in exotic cases
involving symbolic links, though it's possible that the ProjectScope
would also be wrong under similarly exotic circumstances.

Bug: 381384997
Test: existing tests pass; manual verification of bug report
Change-Id: Iae8446a6351dd88f5d86bbe3a648e65a4251cae8

GitOrigin-RevId: 57237fd60acc622815e27367d53de3b9c2280abe