Update DeviceAdapter to key in config more slowly
The super-fast key events of the config may be leading to issues
where the app doesn't process them correctly or in the right order
leading the benchmarking task to get stuck. This change introduces
a slight delay between keys entered that seems to improve how
often the task completes successfully.

Bug: N/A
Test: DeviceAdapterTest.kt
Change-Id: Iaeb94a0a75b188e50fba1403b1baba397b1affad

GitOrigin-RevId: b073ff1b771fd2183452d324312658672c17293d