Remove log warning for previously collected failure details
...in SyncFailureUsageReporter. This class reports data only for Syncs
but the collection of this value in fact happens both for
Build and Sync. There is no easy way at the place of error collection
to distinguish between Sync and Build, no ExternalSystemTaskId
is passed there, only a project path. This resulted in warning in logs
for of any build failures as value is not cleaned up on build start.
This logging is unnecessary as it is in fact impossible to get several
different values for a single build (sync) here. Moreover, reporting of
a stale value remained from previous is impossible as it is cleared up
at the beginning of sync.

Bug: N/A
Test: Existing SyncFailureUsageReporterTest
Change-Id: Id197604fdb2452f3e89ca2e4a39d38696119a593

GitOrigin-RevId: b99808d78848533347899a6bd293a3ee750f870c