Do not use default build actions for KMP modules
These actions are not accurate and slow the build process
quite considerably. Not only this, but these can also cause
build failures if they are not registered tasks.
We should only invoke tasks that are registered for every module, and
not guess on the ones we can invoke without breaking build.

Bug: 372445596
Bug: 158184088
Test: updated
Change-Id: I34d99e44d90bfae1a48aee62f07acf930fe447bc

GitOrigin-RevId: aac6f726212e92a6581fa07cf2e0fb9f2890bb0a