Do not add DesignSurfaceActionHandler to disposer
These are short lived objects. They were added to the disposer to ensure
they were disposed at the same time as the surface, however that creates
a separate leak.

This change ensures that the surface is not strongly held by the
provider and that new references are not added to the disposer.

Test: Existing
Fixes: 384562731
Change-Id: Id0a8f50af8784978a6d9363db3869d17c64e0cc0

GitOrigin-RevId: 8a24c84e2403cbef8bb410c43f77100946a09463