K2: Fix old compose compiler registrar compatibility issue
The default KotlinBundledFirCompilerPluginProvider checks the registrar
service based on the string match with the new compose compiler
registrar name. If it contains an old registrar, the string match fails.
As a result, IntelliJ does not substitute the compose compiler plugin.
This commit fixes the issue by adding a class of
KotlinBundledFirCompilerPluginProvider i.e.,
OldComposeCompilerPluginProviderForK2 to handle the old compose compiler
registrar.

Note that this change has impact on compose compiler plugin
registration. The compose compiler plugin used by LiveEdit and
compose-preview tests is an old one. We had to register its extensions
manually when setting up the tests, but this change makes the IDE
correctly substitute the old compose compiler plugin. As a result, we
have to drop the plugin registration for K2 tests (otherwise, it
registers the composable function kinds twice and causes an exception).

Bug: 383580999
Test: Manually tested an app with an old compose compiler
Change-Id: If9e67309dd965531f239fa96c8381d53366e4d10

GitOrigin-RevId: e506a388b5cc553d7e49d8a00797c6f569bfe757