Fixes incorrect encoding in log parsing file
LogFile was using RandomAccessFile that does not handle unicode
characters. When using it to parse a log file with unicode characters
in them it will return them incorrectly.

Particularly plugin IDs on IJ CE for language packs have unicode
in them. One incorrectly read one was being parsed as a new line,
and not matching the regular expression .* in the tests.

This change properly parses the log as UTF8.

Test: this will be tested by the E2E tests running against IJ CE
Bug: 356400525
Change-Id: I6b1d8102876843e5745d78a20682a29c9fc3e9a2

GitOrigin-RevId: 39b28c58c50bd2fac78f5844e39c8cb225b161b0