[android][gradle] IDEA-363330 disable version catalog accessor in `intellij.android.projectSystem.gradle` module
GradleDslVersionCatalogHandler from `intellij.android.projectSystem.gradle` is a copy of the class with the same name from `intellij.gradle.analysis` module. According to the comment on the copy, it "must be deleted once intellij.gradle.analysis is enabled in Studio or platform navigation relies gradle.dsl.api module".

At this moment, navigation to version catalogs and resolving references to them in IDEA relies on the `android.gradle.dsl` module. The original GradleDslVersionCatalogHandler class uses version catalogs from GradleVersionCatalogModel from `android.gradle.dsl`. Hence, I assume that we don't need its copy in IDEA anymore. AFAIU, it is used in IDEA only when it runs with activated android plugin. But it doesn't bring any valuable functionality in IDEA that the original class could not provide.

I am unsure if it's okay to delete it, so I just disabled it.

GitOrigin-RevId: 8ba145abaf61d93ae432171ad72b9b3e125fd8b0