[Dsl,.toml] Collect catalogs without determining project directory
- `Project.guessProjectDir()`
- `Project.getBaseDir()` - deprecated
- `BuildModelContext.getBaseDirPath(project)` - uses not recommended `getBasePath()`
All these methods are not reliable and could bring unexpected results.

Before this commit, it was possible to encounter the bug when version
catalogs were not collected correctly. The problem was in unexpected
behavior of `guessProjectDir()` for a composite build with a linked
project located outside the root project directory. I could not find a
reliable alternative method. So, reworked the collecting version
catalogs without determining a base project directory at all.

Probably, it makes sense to apply similar changes in
`GradleVersionCatalogViewImpl` - take a value from the `from` property
as a Virtual file, not as a String.

Fixes: 388616181, IDEA-347852
Test: existing GradleSettingsModelTest.
Change-Id: I34bd8355f451aaf06f66392425b0bccc5a09144f

GitOrigin-RevId: 9b36804925064ed9579fa68dd589c2e73205cfe1