Change the order of actions in tests
The call to dispatchAllEventsInIdeEventQueue can have invocation events
involved in it, and we need to execute those before the project is disposed.
This change moves the call to dispatchAllEventsInIdeEventQueue() before
attempting to close and dispose the project.
There are a lot of test flakiness that is happening around our codebase,
and some could be linked to what we were doing here disposing of the project
earlier.

Bug: 385322415
Test: N/A
Change-Id: I8596587be959561aa0029c0ad3fa22662ecf5bd1

GitOrigin-RevId: 9d68480178bba0725807b207df9b91ca639b1236