Fetch report if not returned in batch call
List response sometimes does not return the sample event ID resulting in
the call to batch api failing with empty ID for errorReport. This is due
to the event not having a stack trace. To fetch the event for this issue
we call the api again. This still results in the event having no stack
trace. To denote this, add a panel to StackTraceConsole that shows the
message

Fixes: 386156944
Test: Added
Change-Id: I78a9b4e81536d02eb6af15d3544c6d7c06dd679d

GitOrigin-RevId: b4d74cdc8ded14a7ef4d3c64e5caee696185eacb