Update test projects in designer-perf-test
Update to kotlin 2.0.0 and make it use toml
Also update to API 35

Delete hardcoded kotlin version in ComposeGradleRule
Fix test,in which error occur due to compilation of the project

In perfgate tests replace usage of androidFacet to mainAndroidFacet.
otherwise there is error "ERROR: ClassFileFinder for Module: '%testName%
.app' holder module requested. This is ambiguous. Falling back to the
main module.
"

Bug: 387966225
Test: //tools/adt/idea/designer-perf-tests:intellij.android.designer-perf-tests_tests
Change-Id: I1cc2626fb9abe21fd8e341abb1c46472b6245224

GitOrigin-RevId: 3e42c396861606c3cb90c3ed54b7b686d56c468a