Add automated test for migrated first launch wizard
There is a minor difference in the navigation button
behaviour of the new and old wizards. The old shows a
'Finish' button on the penultimate step, whereas the
new shows a 'Next' button. I've created an additional
automated test to capture this difference and to allow
us to test both flows while the migration is in
progress.

Bug: 311256502
Test: FirstLaunchNewWizardTest
Change-Id: If2c1511f12187fe65152250fe1913412f704045e

GitOrigin-RevId: 2d908adecc97289ffbe9fd0ee04d43f3657c8c62