Delete getLocalInvoker method
The getLocalInvoker() method was added as a hack to get build flags from the blazerc present at the home directory that's inaccessible by the build-api. We've migrated to using canonical blazercs or checked-in blazercs, so this hack is no longer needed.

This is a small CL as part of the build workflow clean-up. In subsequent CLs, we'll modify the BuildInvoker and BuildSystem interfaces in such a way that we have only one getBuildInvoker method so that the caller don't have to care about which invoker to obtain.

Bug:378707836

Test: n/a
Change-Id: I59d9f4d1f2cfad2043845052a358d229497d1fa4

GitOrigin-RevId: 287006495a4412c6db1c41917dbbd5565302c5e5