Fix flaky VitalsTabTest
VitalsTabTest emits an event for the visibility of the tab that may or
may not show up before the initial state is consumed leading to a race
condition. Consume all the events when consuming the initial state to
ensure that the initial state is consumed.

Fixes: 388928988
Test: Enabled

Presubmit-runs-per-test: //tools/adt/idea/app-quality-insights/play-vitals/view:intellij.android.app-quality-insights.play-vitals.view.tests_tests@200

Change-Id: Ife19e08e0a3948534bfaa617c0ff76e76ac11c43

GitOrigin-RevId: 362a164cf5fabc6a63caeeb1aeba8998fabe839a