Cache Results of getClassNames
Use the same `CachedValuesManager` mechanism as
`DefaultClassNameCalculator`.

Also includes a workaround for AndroidX project (see b/388473186) for
details.

The workaround uses a second cache layer where KtFile's that refer to
different instances of the file all map to a canonical representative
instance which is used in the calculation.

The flag is turned off by default.

Fixes: 388473186
Test: Added
Change-Id: Ic8b9cf7867b11d6ed75430d28661f94bfb8305d8

GitOrigin-RevId: 1b37ab6e717066642370352cc71f966e66bd1856