Remove Gradle file handling from AndroidPsiTreeChangeListener
It is (almost) already handled in the listener of GradleFiles.  Extend
the listener's behaviour to update all notifications if either it
detects a new changed file, or if the file containing the change was
relevant but we already had a changed file entry for that file.

The platform debounces updateAll notifications to a granularity of
100ms even for visible editors, so the fact that we might emit
slightly more such notifications is probably OK.

Bug: none filed
Test: existing, manual
Change-Id: I42f30a01fd1faf671a78a1b5b05b15713cd32f66

GitOrigin-RevId: b60faa7be2c95edcb48cb1a49b0c2afaaf998266