Add versions to _ANDROID_STUDIO_SYSTEM_SELECTOR
This will add the major, minor, and micro version to the
-Didea.paths.selector VM arg, so that configs are imported consistently
across versions. Additionally, by adding the micro version we allow for
Android Studio to run simultaneously with versions that have the same
major and minor version.

Previously, the major and minor version were added as part of the build
but that would create an undesired format after adding the micro version
when being stamped (e.g. AndroidStudio2.2024.3 instead of
AndroidStudio2024.3.2).

Bug: 389158350
Change-Id: I12e89a2d282a54d0e875f8737a59559f71f4d8bb

GitOrigin-RevId: 884c14c5a2a8848c32057797ece31d99886aa316