New GradleFiles PsiListener strategy
Don't try to monitor what file is being edited in selected editors and
activate/deactivate a Psi listener on that basis.  Instead, make the
Psi listener be always active, a project service, and filter out
events that are not relevant.

This is not a complete solution to all issues relating to our build
file change detection, though I believe it works approximately as well
as the existing one, and should not leak a PsiListener ever time the
user switches from an editor tab open to a build file.

Bug: 399868758
Test: existing
Change-Id: I0425984cd15a2d29ad2d85609837354f91893b48

GitOrigin-RevId: dc74fe6ee124103ee5c63d7867228e99a9785175