Disable assertions about slow operation running in EDT
With IntelliJ 2024.3 several violations of slow operations in EDT
have been detected. Despite many of them are coming from the platform,
there are few introduced by us or as consequence of platform changes,
since there are no validation in place during development flow.
However, those are going to be tackle one by one and until then the
errors are suppressed linked to the associated bug.

Bug: 390591293
Bug: 391099838
Bug: 391098349
Bug: 391099539
Bug: 391099536
Bug: 391099534
Bug: 391098343
Test: Manual
Change-Id: I3b525ca8cb4e04782d5c8b17fc36372df73a6c82

GitOrigin-RevId: 3ded69d0feb8ae75ee453dd0f59587e0df9015f2