Even more correct GradleFiles tests
Modifying files that exist in the project is likely to be a more
ecologically valid test than creating an empty file and adding to it.

Strictly speaking, this reduces the (mostly accidental) test coverage
of the effect of creating a new "relevant" file.  I'm not sure that
the behaviour of the system under file creation was carefully
considered; future work will define the expected behaviour under VFS
changes (independent of Psi) more carefully, and test for it.

Bug: 345518607
Test: this is it
Change-Id: I64558e96a9f8a4256dcf6332147f0ede2b4454e8

GitOrigin-RevId: 3a51d689b0fffd316646fb7d6048fc2e8a4cf2e2