[Dsl,.toml] support version catalogs in included builds without a settings file.
Before this commit, version catalogs in an included build of a composite build were supported only if the included build had a settings file. If an included build didn't have its own settings file, BuildModel used a settings file from upper directories. This would be correct if the build was a subproject, but incorrect for an included build of a composite build.

There is a difference between a "subproject" and an "included build of a composite build":
1) In the root settings file:
- `include("subproject")` registers subproject of a multi-project build.
- `includeBuild("includedBuild")` registers an included build of a composite build.
2) A subproject uses version catalogs from the root settings file. An included build doesn't because it could have its own settings file and its own version catalogs.

Fixes: 391154478, IDEA-361677
Test: covered with tests in the master branch of IDEA ultimate repository
Change-Id: I1122a51f27154bd7d0287ec197ca7452daff0db6

GitOrigin-RevId: 3058cd7f6937876ec345b4bfb57466ea3a17fddc