Fix ParametrizedPreviewTest
testUiCheckForParametrizedPreview is sometimes failing because of a
timing issue with preview.onActivate(). This method triggers a refresh
asynchronously, but if that takes too long, the refresh completion
listener will already be added and refreshCompled will be updated at the
end of that first refresh instead of the one coming from setting the UI
check mode.
This ensures that we wait for the refresh of preview.onActivate() to
have completed before pursuing the test.

Fixes: 385137823
Test: test code
Change-Id: I621ce5fa0bbd00a53cfae502dbe24a4cf475efc2

GitOrigin-RevId: 28e4e3f4e28da9c66f208b97fe503721bf68b723