Require light inner classes to have light parents
This makes sense in that all of our light inner classes should have
container classes that are also light. From a code perspective, this
change helps with nullability in some upcoming changes I'm making; if we
know the containing class is AndroidLightClassBase, we can write code
that knows `containingClass` and `containingFile` are always non-null.

Bug: n/a
Test: Covered by existing
Change-Id: I27b699b927e1f26901c5fc4e034f60a9997c13c1

GitOrigin-RevId: 1196503c323a0dd71ca0b1dd55de6b2c10822f91