Update DevStatus to transition to server flag setup
Read the deprecation information from ServerFlags for the requested
service. Determines the status of the service depending on the
availability of the flag. If the flag is missing, it is assumed that the
service is still supported. Availability of the flag determines that the
service is marked deprecated by us.

Bug: 389952923
Test: Updated
Change-Id: I3c5afb112c3d2a4e81dd679fb0fcda4d31c517ba

GitOrigin-RevId: 25cc2c7e319c12b1e8c1d6b2ba04c1104a60b971