Set moduleInfo on light class construction
The existing pattern is that derived classes are supposed to call
`setModuleInfo` during their initialization. Most do, but some have
missed it, and this results in the Kotlin plugin not being able to
appropriately identify the classes.

I'm changing the code pattern to require `ModuleInfo` to be passed in
via the constructor. The parameter is currently nullable in the super AndroidLightClassBase for the cases that don't have module information yet, so this change shouldn't affect any behavior. I'll follow up with another change to actually set the module information.

Bug: n/a
Test: covered by existing
Change-Id: I2358ac1cd475880d76310df6b56fe04d64173017

GitOrigin-RevId: ea81c6595a849c827b6a3f15423c3495ed575e0e