Fix module context for lint running in TOML files
This CL improves the code which tries to find an associated
Android module (used when you're editing a file in the editor,
and we want to attach a "main" module for context (like
minSdkVersion). It used to compute *all* android facets in
the project, and then just pick the first one that is an
app module. It now just searches directly and ends the search
when it finds one instead of computing a full list.

More importantly, it also handles the case where there is no
Android module dependency; this happens when you're editing
TOML files.

Test: Existing
Bug: None
Change-Id: I99f46cd44cd88070976ce2a05b1f1d1f63b5a72c

GitOrigin-RevId: 1de7c6615aeced2517b5286290ca59eafb0527e1