Expose AS SDK for Intellij Platform Gradle Plugin
Building the Protocol Buffers plugin from sources uses the IntelliJ
Platform Gradle Plugin, which requires an SDK to build against. The
plugin is hardcoded to expect files to be laid out under the
`android-studio` directory, so this rule exposes the existing files in
an archive with that structure.

The original `-full-linux` was added for a test target with ag/12980024
but is no longer used, so I'm just repurposing it with a slightly
different directory structure and reduced visibility.

Bug: n/a
Test: It builds!
Change-Id: I17b1e466a81357faccdf65b3e3f16c2c0ec3db87

GitOrigin-RevId: ebee35c9568db23faec682644687823d1b99989d