16 KB align metrics for deployment and APK viewer
Record metrics when an APK with native libraries is deployed or when one
is viewed in APK analyzer.

Metrics are recorded regardless of whether the feature server flag is
enabled. However, the bubble-related metrics only fire when a bubble is
shown, and that _is_ gated by the server flag.

The deployment-related metrics also record some device properties so
that we can segment reports by information in them:
- ro.product.cpu.abilist -- for example, 'x86_64,arm64-v8a'
- ro.build.characteristics -- for example, 'emulator,nosdcard,watch'

Bug: 382091362
Test: Existing pass

Change-Id: Id7a38891e76f85b0152d56dc7a08d000fc5c7d87

GitOrigin-RevId: 8da1b55ef774825aad14ab4d14e86a05c986247a