Merge Local Blaze and Bazel Invokers
The Blaze and Bazel invokers are pretty much the same, except the binary type and the binary path. Merge them into a single `LocalInvoker` to avoid duplicate code as part of the cleanup.

Prior to cleanup, they were two different classes, but shared the same command runner and the build result helper.

Bug:374906681

Test: n/a
Change-Id: I7a9a570c48191ba6b05e1b5d227fc8a97ebe9398

GitOrigin-RevId: ed03cdd2727d65a6c6f3258e4d5fe44ebb6febd5