Fix scrolling regression
This ensures that SceneManager#update is called after the render when a
scroll happens within the Views.
We used to have a requestRenderAndLayout method that did this but after
the method was removed, this functionality regressed. We fix it here by
adding a utility method that replicates the render and then forces a
update call that updates the SceneComponent coordinates.

Fixes: 392934332
Test: ScrollInteractionTest
Change-Id: I55bb85d8a0862c9c276cc7815cbea4c1624b0b6f

GitOrigin-RevId: f56823f2b2506f0b1a02b8e791f5acc1f896a5d1