Fix interactive preview zoom-to-fit
When entering interactive Preview a second time, zoom-to-fit wasn't
applied. A system of bitwise keys is used to understand when both
DesignSurface and Preview are done on calculating the right sizes to
apply a correct zoom-to-fit.

When a PreviewMode changes these bitwise integer gets reset. However,
DesignSurface doesn't call resize for certain PreviewModes which is
significant to calculate when applying zoom-to-fit:
* Default: it resizes only when entering Focus and Animation
  Inspection.
* Ui Check: Resizes only when entering from Default Mode and the
  problem panel is closed.
* Animation Inspection: always resizes on enter.
* Focus: always resize on enter.
* Interactive Preview: Resizes only when entering from Focus mode.

This change ensures that calling resetZoomNotifier prevents the
DesignSurface resize notification from being processed when entering
Interactive Preview.

Bug: 381293471
Test: ComposePreviewRepresentationGradleTest
Change-Id: I288ea0e5028c82fb0068808e635f61c03ede0ab5

GitOrigin-RevId: bf7061db744a29aa990f6615d53fac5929870bd8