Fix IndexNotReadyException in Compose preview
The new code to handle KMP previews creates a new path where even if the
model is not ready (useCompose is not available yet) we can look for
existing Preview annotation.
This causes some code that was not running before in non-smart mode to
run now.

This change handles those cases and handles the IndexNotReadyException.
The code can still run but there is information that will not be
available in those cases.

Test: Existing and manually verified in trembling non-smart mode
Bug: 394295660
Change-Id: I5a583805d3d64a4609a7b994d92aa77347deeb05

GitOrigin-RevId: 1c8f38e408f296e8b74f69ed3d5237bb0eaa62f9