Fix VisualizationTest bleak test
openAndCloseVisualizationTool opens a Java file, which creates a
disposable registered to the project. This has nothing to do with the
Validation panel, but this is detected by Bleak as a leak.
To workaround this problem, this replaces opening a Java file with
opening a non-layout xml file, which tests the same behaviour.

Fixes: 390984556
Test: VisualizationTest
Change-Id: I5e247f7692bab9f8d9b6ff1cec038e711789be59

GitOrigin-RevId: 5725fb24a13204655292db3e7ebeb9678d96f4fc