Remove dead code
The change in I6ef47d24a09ce2f2f2faaaf3d2bd439d01437b9f inadvertently
changed the effective order of tests, only testing whether external
build files were modified after testing whether any build file at all
was modified, and so meaning that the specific "External build
files have changed..." banner was never seen (and also that the
special "force external native build model generation" flag was never
applied to sync).

Since (a) the whole GradleFiles change detection mechanism is
substantially lossy in the first place (b) to my knowledge we
haven't received problem reports relating to this in ~18 months
and (c) this kind of special-casing is incompatible with the
platform's AutoImport mechanism (which we want to move to), delete the
dead code rather than attempting to restore the functionality, such as
it was.

Bug: none filed
Test: existing test also deleted
Change-Id: Ibdd79d46794541059154039278df2340ee4344c4

GitOrigin-RevId: fdd1e2a42769d2ad6d9a254142377dcc2c4de4e7