Disable rundebug.use.pm.terminate
Ceding control of stopping the app to the package manager causes IWI
deployments to never stop the app, as they do not engage with the
package manager except in cases of fallback deployments.

IWI deployments *may* function regardless, but this is a coincidence due
to a race condition caused by the presence of force-stop in
AndroidProcessHandler; if the deployment finishes before the
ProcessHandler force-stops the app, then the bug occurs regardless.

Bug: 389067070
Test: Manual
Change-Id: Ica379e912b4cc460646bda196db126246e300626

GitOrigin-RevId: 262b408aa7663924a694b418ded0b688f23e507b