Integration E2E test to Gemini chat functionality
* The test reuses mock auth server from
com.google.gct.login2.LoginE2ETest test and the logic of logging in
* The test also implements a fake RPC server that mocks the Aida model
that provides completion/code generation/chat conversations
* In order to leave the production AidaEndpoint logic of connecting to
the endpoint untouched we had to create a temp truststore and configure
Android Studio instance being tested to use it to bypass the gRPC chanel
SSL/TLS checks. (It's a bit hacky)
* The test performs google authentication, performs Gemini onboarding
and sends a text to the chat asserting the response content.

Bug: 394033793
Test: GeminiChatTest
Change-Id: I78953403c42b76575c0f6201084058860b1a0070

GitOrigin-RevId: 88f62f260369ee938e77bf6ec3834c12960e5039