Refactor MaterialVdIconsLoader
This reduces the memory usage by using a single model and avoiding
copies.
The idea is to move to a model where the Model implementation notifies
subscribers about changes. For now, we keep the same logic, but we
refactored the internal to not make additional copies.

The change also moves the sorting responsibility to the UI and not the
internal model.

Bug: 374247114
Test: IconPickerDialogTest
Change-Id: I60b226c7ca0f0e7c1c9ecc0d215a7faf2b890447

GitOrigin-RevId: ae915c97c70832dd056ee682b6a5c133bf284822