Somewhat de-generalize core DependencyManagement
Do not claim to work with arbitrary GradleCoordinates any more:
instead, restrict our domain to the set of artifacts encoded in
GoogleMavenArtifactId (and, specifically, do not allow requesting
specific versions from client code).

This is not a complete solution to the problem that the
DependencyManagement offers interfaces to functionality that it cannot
in general provide, but it's a start, and in particular it removes a
counterintuitive use of GradleCoordinate in the Blaze project system.

Bug: 279149611
Bug: 279886738
Test: existing test adjusted
Change-Id: I40826a43a36818dc6e03fbc6894ed54fb7b3511e

GitOrigin-RevId: aa2d0be4006635337f6477fd20c27781a71f1ec3