Avoid generating preview for Maven import action
The Maven import action modifies both the current file and the Gradle
dependencies file, and using WriteCommands to do so. This doesn't play
well with the default implementation of IntentionAction.generatePreview,
which assumes it can run that logic on a background thread on a fake PSI
file.

Preview doesn't really make sense in this case anyway, since the
modifications are at a code location other than where the red squiggle
is, so I'm addressing this by just avoiding a preview for this action.

Fixes: 396483011
Test: Manually verified
Change-Id: I74d7546b875935503285c074956bb27a5f44dd47

GitOrigin-RevId: b1274fe7452d8ea363b03d8a18cc7ce179962673