Build dependencies in the same way
regardless of how the targets to build were collected, i.e. build
whole project as if all its targets were selected manually.

Note that this change removes special handling of aliases as they have
not been handled in the non-whole-project case. This may result in code
analysis not being enabled in some targets even after enabling analysis
in the whole project. This will be fixed for all builds by transitioning
to a simplified project graph that only tracks targets of known kinds.

Bug: n/a
Test: existing
Change-Id: Ifae478999feb2b7ea4f98b3a38818cef102891e1

GitOrigin-RevId: 8e4c6fbad8f5bb434d37eac7364fa819dd00e170