Add Declarative WatchFace configuration executor
The Declarative WatchFace configuration is different from the other Wear
Configurations in that it doesn't have a main component to run.

The existing AndroidWearConfigurationExecutor and
AndroidConfigurationExecutorBase rely on an AppRunSettings being present
. This AppRunSettings requires a ComponentLaunchOptions which relies on
the ComponentType to be run.

In the case of declarative watch faces, we don't launch a specific
component. Instead, we install the apk and then launch a watchFaceId
corresponding to the application ID.

For this reason, this CL goes with the approach of having a separate
executor.

Bug: 392849344
Test: AndroidDeclarativeWatchFaceConfigurationExecutorTest
Change-Id: I92f075aa386a0d3beccbfbb1ffa7901ee11cd33a

GitOrigin-RevId: 4ca4711e53b1263e36708cdfa7790c205916d318