Rewrite mlkit dependency management to use Gradle Project System
It's pretty embedded in the mlkit module for now, so use that fact to
provide specialized handling of the version-ful dependencies that
mlkit specifies.

(I actually think that none of this is currently functional at all,
but this allows breaking the dependency to DependencyManagement and
consequent simplification there, so it's worth it anyway.)

Accept cosmetic changes to the Add Dependency dialog message along the
way.

Bug: 149224613
Bug: 279149611
Bug: 279886738
Test: existing test adjusted; manual test of add dependency
Change-Id: I811bde91713faed81f7e77f678218f28f5023609

GitOrigin-RevId: 90a17f4cbb15760ca195cef05714f1818a44eade