New interface for registered dependencies
Construct a new interface for querying and otherwise dealing with
registered dependencies.  Tease out different interfaces for return
types, and implement them for Gradle and for the fake-ish
TestProjectSystem.

Rename analyzeDependencyCompatibility to
analyzeCoordinateCompatibility in order to be able to use the
analyzeDependency- name with the RegisteredDependencyId interfaces.

Use the new interface in Module.addDependenciesWithUiConfirmation

Bug: 149224613
Bug: 279149611
Bug: 279886738
Bug: 369433182
Test: existing DependencyManagement tests
Change-Id: Ib3a0dbd44c1dacdf9846ed608d66b05007f39870

GitOrigin-RevId: 85b540ed9bdb2f5e6e422686f123d8d836ab474a