Fix and re-enable LanguageHighlightingTest
One of the highlights this test checks for comes from an Android lint
check. It turns out that Android lint runs *after* the normal
highlighting pass, and the platform API we're calling doesn't actually
guarantee that it has finished.

This change adds code that specifically runs all code analyzer passes
before collecting highlight information. This is comparable to what the
standard CodeInsightTestFixtureImpl does when it collects highlight
information.

2000x run (Windows):
https://fusion2.corp.google.com/invocations/53074156-9c77-4c59-a6ff-3f43fb81720c

2000x run (Linux):
https://fusion2.corp.google.com/invocations/1c269a32-5072-4e07-a61a-78ff9f92fbc6

Fixes: 396863968
Test: //tools/adt/idea/android/integration:LanguageHighlightingTest_windows
Change-Id: If95a9a38943d8bb28021a42bdbd1483b3c3ae6d0

GitOrigin-RevId: 23403e9fef29365c2fbc74005c74a0a067305542