Synchronize framework ids initialization
ResourceIdManagerBase is initialized once per module. Since this can
happen in parallel and in multiple threads there is change that the
framework id initialization flow is triggered multiple times.
This can lead to crashses since the code is not thread safe.

The change synchronizes the initialization for the ids and also makes
some minor changes to allow for testing this behaviour.

Fixes: 397431390
Test: ResourceIdManagerBaseTest
Change-Id: Ifcdfd1f94721adb06648b5ee36bc9c40591017ee

GitOrigin-RevId: 3c07c2d9a95a4c9ade29939cd47f53a696bd858c