Implement on-device rendering of hovered and visible nodes
This change adds Studio support for rendering of hovered nodes and
visible rectangles.

Visible rectangles are all the rectangles visible on screen, Studio
sends them over each time there is a model change.

Hovered rectangles are the rectangles currently being hovered. The
on-device rendering detects the hover coordinates, sends them to
Studio, which checks the model and responds with the draw instructions
for the view node at those coordinates.

This change introduces a new OnDeviceRendererModel, which holds the
state of on-device rendering.

This model is currently only used for on-device rendering, but once
standalone Layout Inspector is deprecated should become the only
render model used by embedded Layout Inspector.

Test: OnDeviceRenderingClientTest
Test: OnDeviceRendererModelTest
Test: OnDeviceRendererPanelTest
Test: RootRendererPnalTest
Bug: 398201661
Change-Id: Id4394d89a13eb516957f9526cceba37116832b1a

GitOrigin-RevId: 31ace18ee93c5459f663284d0b432e6107ae23ae