Add support for right click in on-device rendering
When the user right clicks a component in Layout Inspector, they
expect to see a context menu appear. In order to implement this for
on-device rendering, we need to detect the right click on the device.

Once the right clicks is detected, the coordinates of view that was
clicked are sent to Studio, which shows the popup menu for that view
on the last known mouse coordinate.

Currently this will not work on emulators, where right clicking on
an app causes the app to close. This does not happen on XR emulators
and physical devices. This inconsistency is reported in b/388871785.

Test: OnDeviceRenderingClientTest
Test: OnDeviceRendererModelTest
Test: OnDevieRendererPanelTest
Bug: 398201661
Change-Id: I1513ae5763d9e2de546d1d2535d1791cd1b9b662

GitOrigin-RevId: 3c899987f2ebb903869df22b679b7d5ad36acd7a