Catch UastAnnotationAttributesProvider IndexNotReadyException
It looks like there are a few paths where we try annotation resolution
while not in smart mode. Since the preview should only work in smart
mode, this should not happen but there might be some corner cases like
the editor instantiation (where we look for the available previews)
where this might be happening.

This CL handles the case and logs a warning since we should handle this
at a higher level.

Bug: 398265392
Test: UastAnnotationAttributesProviderTest
Change-Id: I0f2d19620f5bbf6a454c5c29e736b132f2fcdaad

GitOrigin-RevId: bc28e147119ed00abf3cbb0be752e229ed420604