Fix JavaLanguageLevelDeprecationOutputParserTest with IntelliJ 2025.1
With IntelliJ 2025.1 a new reporter parser was added called
GradleCompilationReportParser which is taken into consideration on the
BuildOutputInstantReaderImpl when parsing compilation exceptions being
able to properly parse them and trigger new build events, despite that
those were always present but ignored previously.

However, the created new MessageEvent build event by that parser
doesn't contain consistent parentId with the task that actually
failed the compilation in the attached case must be
:app:compileDebugJavaWithJavac but due race condition end up pointing
to potentially different task. This is being tracked on IDEA-368488,
in the meantime build output events are filtering by expected type.

Bug: 398156145
Test: JavaLanguageLevelDeprecationOutputParserTest
Change-Id: I8ab838bc2f51ab445bcedca13bc01fde1b202c36

GitOrigin-RevId: d7fec867fc289e02188ae6294e3a272c4c97674c