Show loading text when checking for existing icons
The validator has been updated to show a loading icon with the text
"Checking for existing icons..." while the check is ongoing. The
'Next' button of the step is also disabled until the check is complete
to ensure the user sees any warnings before they proceed to the next
step.

Bug: 390731020
Test: Manual
Change-Id: Ic01624c54667334ff72b8a76a1a28dce9009ce16

GitOrigin-RevId: 9ac356e140744a9bdb8ee524c52d147c23ff94b9