Update RenderingBuildStatusManager to handle Fast Preview
When handling the initial build status, RenderingBuildStatusManager was
never looking for classes that have been compiled via Fast Preview. This
could cause some "Out of date" cases when, for example, a new file is
created and built via Fast Preview but not via Gradle. The preview would
consider that class as non-existent and then show the wrong status.

Fixes: 398061031
Test: RenderingBuildStatusManagerResourcesTest,
 RenderingBuildStatusManagerTest, ModuleClassLoaderOverlaysTest
Change-Id: Id7c95b8b8e431befb462709e3bbf7f7b44610cba

GitOrigin-RevId: b9371403eb2c4acd15dfc7e5754d606c9ab21cfa