Remove slow call from ResourceNotificationManager#addListener
addListener was making a call to the
StudioResourceRepositoryManager.getProjectResource to ensure that
listeners were registered in a particular order. addListener should be a
fast call so this CL moves the listeners to a two tier structure where
the resource managers get to subscribe to an "early" notification topic
to achieve the same result without making addListener heavy.

Test: ResourceFolderManagerTest and other resource tests
Fixes: 398925803
Change-Id: Iae7464317de49dac3e38822bbebb67f98456c329

GitOrigin-RevId: 92545918fb4bc68f6ad874d7ae42c8ea4c94f1f5