Remove unused annotationFilter parameter
This parameter was always set to "{ true }" and was only set within a
unit test.

This lambda parameter is used as a field in CachedValuesKey which is not
currently a problem as it refers to the same lambda instance each time.
However, it's best to remove it from the cache key because it is
error-prone. If someone sets a lambda on the fly later (instead of val
for example), the lamba hashcode/equals will no longer be equal,
leading to the cache key being different on each call, like it's
currently the case for the toValues lambda.

Bug: 397205975
Test: existing tests should still pass
Change-Id: If55b5c5d5acf700522af410e50a4b21c7aded7aa

GitOrigin-RevId: 84c86c90f5ffb0436f8f10144ca5d7eccd541df2