Maven artifact lookup to include more available versions
Until now this query only aimed at getting the latest version. By adding
"core=gav" to the lookup URL we get a list of available versions.

With this approach the existing result limit of 50 can be low. Since the
maximum allowed result limit appears to be 200 - this will be used. No
significant time increase noticed.

One downside of this solution is that a more generic search (where no
artifact name is given) can bring a limited list of versions with no
indication that the list is incomplete. For example a search for
'com.google.auth' would yield 6 artifacts, amongst which the
'google-auth-library-appengine' one will offer only 41 versions ranging
from  1.33.1 down to 1.3.0. When a more precise search is attempted
'com.google.auth:google-auth-library-appengine' the result table
contains all the 89 versions down to 0.2.0.

Fixes: 379985077
Test: MavenCentralRepositoryTest
Change-Id: Ia4529cf4da4cbe4ef8d56e89fd60c9ee757bca2c

GitOrigin-RevId: 5d526cae6546068f7106a59b9d09614c3be2f060