Modify signature of getBuildInvoker
The context passed while creating a BuildInvoker object is no longer
used by the invoker since the new methods to invoke blaze/bazel are in
place. This change removes this unused context parameter.

As a result, the context needs to be passed to getBlazeInfo method,
which also makes it consistent with other methods that invoke blaze and
bazel.

Bug:374906681

Test: Existing tests
Change-Id: Ia7bcf7e012bdcffdba747fa78b58766101dd5b19

GitOrigin-RevId: 4893fdcf379a09328805382720c05e76948ca9f6