Support package references from (test)Namespace in Declarative
This involves us making DeclarativeLiteral be a
ContributedReferenceHost, with the minimal changes to the Grammar-Kit
bnf and PsiImplUtil, and somewhat more to the generated files (mostly
removed, but there seems to be some version skew somewhere).  Once
that's done, the references can be generated essentially just as with
the other two implementations.  Rename the file (though not any
individual Contributor or Provider) to a language-neutral name.

Mind you, Gradle Declarative does not yet support the testNamespace
property.

Bug: 214570613
Test: manual
Change-Id: Ia9597337e629cfe94861343551e3627fc6dfff5c

GitOrigin-RevId: 9340933ccc079df91417100114d7d07adbe1ae57