More robust way to disable Kotlin highlighting
The previous logic broke with IntelliJ 2025.1, where the Kotlin
highlighter was split into several different classes with no publicly
accessible base class.

This change makes use of a couple internal APIs, although it is
slightly better than the alternative JetBrains/android commit
https://github.com/JetBrains/android/commit/fed7b736cc
which relies on reflection (and happens to have a bug).

Change-Id: I9bd8e3f721fb50ac971aa4db21e5edf6333460fe

GitOrigin-RevId: 74ffe6d5aa200b50dd9b5bae61b4ef90ba05cd64