Turn on indexing for an indexing target
Or "turn off skipping index initialization".
ConfigurationCacheFailureTest is in a target apparently named to have
indexing enabled, but by some oversight (possibly the double negative)
index initialization was still skipped for it.  Without index
initialization, the test passes ~99% of the time, but occasionally
test teardown apparently has a race with index initialization (from
some other part of the platform, presumably) and we get an NPE from
HeavyPlatformTestCase.

Bug: 399167593
Test: existing
Change-Id: Icafd83012076f6201116d85b9069cce93b089651

GitOrigin-RevId: ec175bcc0dc9b728c2d7d527c2445d20b3926abb